pyproject.toml
[project]
name = "biz-assistant"
version = "0.1.0"
description = "Biz Licensing Assistant - FastAPI + Streamlit skeleton"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pdfplumber>=0.11.7",
    "pydantic>=2.11.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "uvicorn[standard]>=0.35.0",
]
